Fix bug of Doctrine MongoDB support
