+++ Resources/config/mongodb.xml
@@ -13,6 +13,12 @@
             <argument>%fos_user.model.user.class%</argument>
         </service>
 
+        <service id="sonata.user.manager.user" class="Sonata\UserBundle\Model\UserManagerProxy">
+            <argument>Sonata\UserBundle\Document\User</argument>
+            <argument type="service" id="doctrine_mongodb" />
+            <argument type="service" id="sonata.user.mongodb.user_manager" />
+        </service>
+
         <service id="sonata.user.mongodb.group_manager" class="Sonata\UserBundle\Document\GroupManager" public="false">
             <argument type="service" id="fos_user.document_manager" />
             <argument>%fos_user.model.group.class%</argument>
