+++ Resources/config/orm.xml
@@ -13,7 +13,7 @@
             <argument>%fos_user.model.user.class%</argument>
         </service>
 
-        <service id="sonata.user.manager.user" class="Sonata\UserBundle\Entity\UserManagerProxy">
+        <service id="sonata.user.manager.user" class="Sonata\UserBundle\Model\UserManagerProxy">
             <argument>Sonata\UserBundle\Entity\User</argument>
             <argument type="service" id="doctrine" />
             <argument type="service" id="sonata.user.orm.user_manager" />
